==> archives/06-Exploration-Title-Principals.ipynb <==
# Dataset Exploration - Title (Principals)

This notebook will assist with exploring the data set to understand the data that it contains.

import pandas as pd

# load the title basics dataset
df = pd.read_csv('./data/title.principals.tsv', index_col=None, na_values=['NA'],
sep='\t',low_memory=False)

# display row and column counts
df.shape

# display column names
df.columns.tolist()

from IPython.display import display

# load the field descriptions
fld_dicitonary = pd.read_csv("./data/title_principals_flddesc.csv", index_col=0, sep='^')
pd.set_option('display.max_colwidth', -1)

# display descriptions
fld_dicitonary

# display the first 25 rows
df.head(25)

df.ordering.unique()

# display the last 25 rows
df.tail(25)
