==> archives/09-Data-Clean-Title-Basics.ipynb <==
# Data Clean on IMDB Title Basics Data Set

# import libraries
import pandas as pd
import pickle

# load the title basics dataset
tBasics = pd.read_csv('./data/title.basics.tsv', sep='\t',low_memory=False)

# display row & column counts
tBasics.shape

# view data frame
tBasics.head()

# drop all adult titles
tBasicsSub = tBasics[tBasics.isAdult == 0]

# display row & column counts
tBasicsSub.shape

# drop all title types except movies
tBasicsSub = tBasicsSub[tBasicsSub.titleType == 'movie']

# display row & column counts
tBasicsSub.shape

# view data frame
tBasicsSub.head()

# drop all movies that have no runtime value
tBasicsSub = tBasicsSub[tBasicsSub.runtimeMinutes != '\\N']

# display row & column counts
tBasicsSub.shape

# convert runtime minutes to an integer type
tBasicsSub.runtimeMinutes = pd.to_numeric(tBasicsSub.runtimeMinutes , errors='coerce')

Drop any movie title that has a runtime of 79 minutes or less. The Screen Actor Guild (SAG) defines a feature film at 80 minutes or greater in runtime length. 
https://screenwriting.io/what-is-a-feature-film/
https://en.wikipedia.org/wiki/Feature_film

# select all movies that are 80-minutes or more in runtime minutes
tBasicsSub = tBasicsSub[tBasicsSub.runtimeMinutes > 79]

# display row & column counts
tBasicsSub.shape

# view data frame
tBasicsSub.head()

# save cleaned data set to a pickle file
tBasicsSub.to_pickle('v1_tBasicsCleaned.pkl')
