==> archives/10-Data-Clean-Title-Aka.ipynb <==
# Data Clean on IMDB Title Aka Data Set

# import libraries
import pandas as pd
import pickle

# load the title also known as dataset
tAka = pd.read_csv('./data/title.akas.tsv', quoting=3, sep='\t',low_memory=False)

# display row & column counts
tAka.shape

# view data frame
tAka.head()

# drop all regions except US
tAkaSub = tAka[tAka.region == 'US']

# display row & column counts
tAkaSub.shape

# create a copy of the tAkaSub data frame
tAkaSubUS = tAkaSub.copy()

# display row & column counts
tAkaSubUS.shape

# view data frame
tAkaSubUS.head()

# drop all columns except titleId
tAkaSubUS.drop(['ordering', 'title', 'region', 'language', 'types', 'attributes', 'isOriginalTitle'], axis=1, inplace=True)

# view data frame
tAkaSubUS.head()

# drop duplicate titleId's
tAkaSubUS.drop_duplicates(inplace=True)

tAkaSubUS.shape

# save cleaned data set to a pickle file
tAkaSubUS.to_pickle('v1_tAkaCleaned.pkl')
