==> archives/11-Data-Clean-Title-Ratings.ipynb <==
# Data Clean on IMDB Title Data Set

# import libraries
import pandas as pd
import pickle

# load the title ratings dataset
tRatings = pd.read_csv('./data/title.ratings.tsv', index_col=None, na_values=['NA'],
sep='\t',low_memory=False)

# display row & column counts
tRatings.shape

# convert numVotes to an integer type
tRatings.numVotes = pd.to_numeric(tRatings.numVotes , errors='coerce')

# convert averageRating to a float type
tRatings.averageRating = pd.to_numeric(tRatings.averageRating , errors='coerce')

tRatings.head()

# save cleaned data set to a pickle file
tRatings.to_pickle('v1_tRatingsCleaned.pkl')
