==> archives/12-Data-Clean-Title-Principals.ipynb <==
# Data Clean on IMDB Title Principals Data Set

# import libraries
import pandas as pd
import pickle

# load the title basics dataset
tPrincipals = pd.read_csv('./data/title.principals.tsv', index_col=None, na_values=['NA'], sep='\t',low_memory=False)

tPrincipals.shape

# drop all categories except director, actor, actress and writer
tPrincipalsSub = tPrincipals[((tPrincipals.category == 'director') | (tPrincipals.category == 'actor') |
                              (tPrincipals.category == 'actress') | (tPrincipals.category == 'writer'))]

tPrincipalsSub.shape

tPrincipalsSub.category.value_counts()

# save cleaned data set to a pickle file
tPrincipalsSub.to_pickle('v1_tPrincipalsCleaned.pkl')
