==> archives/13-Data-Clean-Names-Basic.ipynb <==
# Data Clean on IMDB Name Basic Data Set

# import libraries
import pandas as pd
import pickle
from datetime import date

# load the names basics dataset
nBasics = pd.read_csv('./data/name.basics.tsv', index_col=None, na_values=['NA'], sep='\t',low_memory=False)

nBasics.shape

nBasics.head()

# drop all movies that have no runtime value
nBasicsSub = nBasics[nBasics.deathYear == '\\N']

# display row & column counts
nBasicsSub.shape

# drop all movies that have no runtime value
nBasicsSub = nBasicsSub[nBasicsSub.birthYear != '\\N']

# display row & column counts
nBasicsSub.shape

nBasicsSub.head()

nBasicsSub.birthYear.unique()

# convert runtime minutes to an integer type
nBasicsSub.birthYear = pd.to_numeric(nBasicsSub.birthYear , errors='coerce')

nBasicsSub.head()

# drop all movies that have no runtime value
nBasicsSub = nBasicsSub[nBasicsSub.birthYear > 1899]

# display row & column counts
nBasicsSub.shape

def calculate_age(born):
    today = date.today()
    return today.year - born

nBasicsSub['Age'] = calculate_age(nBasicsSub.birthYear)

# display row & column counts
nBasicsSub.shape

nBasicsSub.head()

nBasicsSub.Age.unique()

# drop all movies that have no runtime value
nBasicsSub = nBasicsSub[nBasicsSub.Age < 103]

# display row & column counts
nBasicsSub.shape

# save cleaned data set to a pickle file
nBasicsSub.to_pickle('v1_nBasicsCleaned.pkl')
