==> archives/14-Data-Merge-Creating-Movie-Selection-Data-Set.ipynb <==
# Data Merge Creating Movie Selections Data Set

import pandas as pd

tBasicsCleaned = pd.read_pickle('v1_tBasicsCleaned.pkl')

tBasicsCleaned.shape

tAkaCleaned = pd.read_pickle('v1_tAkaCleaned.pkl')

tAkaCleaned.shape

tRatingsCleaned = pd.read_pickle('v1_tRatingsCleaned.pkl')

tRatingsCleaned.shape

### Merging Aka Into Basics

movieSelections = tBasicsCleaned.merge(tAkaCleaned, left_on='tconst', right_on='titleId', how='inner')

movieSelections.shape

### Merging Ratings Into Movie Selections

movieSelections = movieSelections.merge(tRatingsCleaned, left_on='tconst', right_on='tconst', how='inner')

movieSelections.shape

# save cleaned data set to a pickle file
movieSelections.to_pickle('v1_movieSelections.pkl')
