==> archives/15-Data-Merge-Creating-Cast-Crew-Selection-Data-Set.ipynb <==
# Data Merge Creating Cast & Crew Selection Data Set

import pandas as pd

tPrincipalsCleaned = pd.read_pickle('v1_tPrincipalsCleaned.pkl')

tPrincipalsCleaned.shape

nBasicsCleaned = pd.read_pickle('v1_nBasicsCleaned.pkl')

nBasicsCleaned.shape

movieSelections = pd.read_pickle('v1_movieSelections5kVotes.pkl')

# 08/09 scope change
#movieSelections = pd.read_pickle('v1_movieSelections.pkl')

movieSelections.shape

### Merge Name Basics Into Principals

castCrewSelections = tPrincipalsCleaned.merge(nBasicsCleaned, left_on='nconst', right_on='nconst', how='inner')

castCrewSelections.shape

### Merge Movie Selections Into Cast & Crew Selections

castCrewSelections = castCrewSelections.merge(movieSelections, left_on='tconst', right_on='tconst', how='inner')

castCrewSelections.shape

castCrewSelections.category.value_counts()

# save cleaned data set to a pickle file
castCrewSelections.to_pickle('v1_castCrewSelections.pkl')
