==> archives/16-Movie-Selection-EDA.ipynb <==
# Movie Selection Exploratory Data Analysis (EDA)

import pandas as pd

movieSelections = pd.read_pickle('v1_movieSelections.pkl')

movieSelections.shape

# select all movie titles with 5000 or more votes
movieSelections5kVotes = movieSelections[movieSelections.numVotes >= 5000]

# 08/09 - scope change
#movieSelections95Pct = movieSelections[movieSelections.numVotes >= 17149] # 95% percentile

movieSelections5kVotes.shape

# save cleaned data set to a pickle file
movieSelections5kVotes.to_pickle('v1_movieSelections5kVotes.pkl')

# 08/09 - scope change
#movieSelections95Pct = movieSelections95Pct.sort_values(by='averageRating', ascending=False)

### Top 10 Ranked Movies By Average Rating

# 08/09 - scope change
#movieSelections95Pct.head(10)

### Top 10 Ranked Movies By Number of Votes

top10ByVotes = movieSelections5kVotes.sort_values(by='numVotes', ascending=False)

top10ByVotes.head(10)
