==> archives/17-Cast-Crew-Selection-EDA.ipynb <==
# Cast & Crew Selection Exploratory Data Analysis (EDA)

import pandas as pd

castCrewSelections = pd.read_pickle('v1_castCrewSelections.pkl')
display(castCrewSelections.shape)
castCrewSelections.head()

#movieSelections = pd.read_pickle('v1_movieSelections5kVotes.pkl')
#movieSelections.shape

#castCrewSelections = castCrewSelectionsIn.merge(movieSelections, left_on='tconst', right_on='tconst', how='inner')

# display column & row information
#display(castCrewSelections.shape)

# display top 5 rows
#castCrewSelections.head()

actorSelections = castCrewSelections[castCrewSelections.category == 'actor']
actorSelections.shape

actorSelectionsAgg = (actorSelections.groupby('primaryName')
   .agg({'tconst':'count', 'numVotes': 'sum'})
   .reset_index()
   .rename(columns={'Movies Titles':'Average Rating'})
   )
actorSelectionsAgg.shape

actorSelectionsAgg.head()

actorSelectionsAgg = actorSelectionsAgg[actorSelectionsAgg.tconst >= 30]
actorSelectionsAgg.shape

actorSelectionsAgg.sort_values('numVotes', ascending=False, inplace=True)

# Top 10 Actors Ranked by Highest Movie Rating With 10 or More Films 

actorSelectionsAgg.head(10)

actressSelections = castCrewSelections[castCrewSelections.category == 'actress']
actressSelections.shape

actressSelectionsAgg = (actressSelections.groupby('primaryName')
   .agg({'tconst':'count', 'numVotes': 'sum'})
   .reset_index()
   .rename(columns={'Movies Titles':'Average Rating'})
   )
actressSelectionsAgg.shape

actressSelectionsAgg.head()

actressSelectionsAgg = actressSelectionsAgg[actressSelectionsAgg.tconst >= 30]
actressSelectionsAgg.shape

actressSelectionsAgg.sort_values('numVotes', ascending=False, inplace=True)

# Top 10 Actress Ranked by Highest Movie Rating With 10 or More Films 

actressSelectionsAgg.head(10)

directorSelections = castCrewSelections[castCrewSelections.category == 'director']
directorSelections.shape

directorSelectionsAgg = (directorSelections.groupby('primaryName')
   .agg({'tconst':'count', 'numVotes': 'sum'})
   .reset_index()
   .rename(columns={'Movies Titles':'Average Rating'})
   )
directorSelectionsAgg.shape

directorSelectionsAgg.head()

directorSelectionsAgg = directorSelectionsAgg[directorSelectionsAgg.tconst >= 10]
directorSelectionsAgg.shape

directorSelectionsAgg.sort_values('numVotes', ascending=False, inplace=True)

# Top 10 Directors Ranked by Highest Movie Rating With 10 or More Films 

directorSelectionsAgg.head(10)

writerSelections = castCrewSelections[castCrewSelections.category == 'writer']
writerSelections.shape

writerSelectionsAgg = (writerSelections.groupby('primaryName')
   .agg({'tconst':'count', 'numVotes': 'sum'})
   .reset_index()
   .rename(columns={'Movies Titles':'Average Rating'})
   )
writerSelectionsAgg.shape

writerSelectionsAgg.head()

writerSelectionsAgg = writerSelectionsAgg[writerSelectionsAgg.tconst >= 10]
writerSelectionsAgg.shape

writerSelectionsAgg.sort_values('numVotes', ascending=False, inplace=True)

# Top 10 Writers Ranked by Highest Movie Rating With 10 or More Films 

writerSelectionsAgg.head(10)
